moving validation inside capabilities
