+++ src/main/java/com/appium/utils/ArtifactsUploader.java
@@ -74,7 +74,7 @@ public class ArtifactsUploader {
                 .getCapabilityObjectFromKey("windows");
         if (android != null && android.has(app)
             && (platform.equalsIgnoreCase("android")
-            || platform.equalsIgnoreCase("both"))) {
+            || platform.equalsIgnoreCase("all"))) {
             JSONObject androidApp = android.getJSONObject("app");
             String appPath = androidApp.getString("local");
             if (isCloud(hostMachine) && androidApp.has("cloud")) {
@@ -84,7 +84,7 @@ public class ArtifactsUploader {
         }
         if (iOSAppPath != null && iOSAppPath.has("app")
             && (platform.equalsIgnoreCase("ios")
-            || platform.equalsIgnoreCase("both"))) {
+            || platform.equalsIgnoreCase("all"))) {
             if (iOSAppPath.get("app") instanceof JSONObject) {
                 JSONObject iOSApp = iOSAppPath.getJSONObject("app");
                 if (iOSApp.has("simulator")) {
